Fix template check failures
There are issues in the template that we want to ignore: lambda DLQs and
KMS key rotation. The added argument, skip_check, will suppress these
issues.

GitHub complains about the arguments that we pass to the checkov action.
The specific release documents these as valid, hopefully resolving this
warning.

Issue: PLAT-20
Co-authored-by: Dale Peakall <dale.peakall@digital.cabinet-office.gov.uk>