Prevent failed run of preTrafficHook lambda
A surprise: lambdas that are used as hooks in another lambda's
deployment preference hooks MUST have a function name that begins with
'CodeDeployHook_', hence the name for the PreTrafficHook lambda. If this
prefix isn't present, then the lambda is not granted the permissions to
write its validation status back to CodeDeploy.
See https://github.com/aws/serverless-application-model/issues/219

Also includes a fix for the lambda, where an extra JSON parse needed to
be performed.

Issue: PLAT-51
Co-authored-by: Adrian Wong <adrian.wong@digital.cabinet-office.gov.uk>