Updated example for ephemeral builds
When used in normal builds the stack that depends on the parameters
can derive the parameter path by combining the `Environment`
parameter with the parameter specific name.

In ephemeral builds, we also need a `LocalName` that provides a
unique component that can be used when defining and referencing
the parameters.

Overrides of values for ephemeral environments are possible using
parameters that have defaults which are used instead of the
mapping when deploying an ephemeral environment. Even though the
mapping won't be used a value must be defined for the
corresponding environment in the mapping.

Issue: PLAT-68