Include OpenAPI spec in sam-app Serverless::Api
Required to demonstrate how cloudformation includes can be used and be
valid with the pipeline. Also provides a useful test case.

A surprising thing: When the lambda ARNs have to be provided in the
OpenAPI spec 'uri' properties, for HelloWorldFunction, an additional
`:HelloWorldFunction` needed to be put at the end of the ARN in order
for the function to be executable by API Gateway. This was not required
for HelloWorldFunction2. A theory is that the `:HelloWorldFunction` is
the AutoPublishAlias value.

Issue: PLAT-87
Co-authored-by: Mauro Vilasi <mauro.vilasi@digital.cabinet-office.gov.uk>