Add certificate to API Gateway
We added "UseCertificate" as a condition since we don't own a domain per account.
We are not using the domain key for the API resource since we can not use conditionals, therefore, we manually created the base path and the domain name for the API. Normally this wouldn't be required.

Issue: PLAT-112
Co-authored-by: Cristhian Da Silva <cristhian.dasilvafernandez@digital.cabinet-office.gov.uk>
Co-authored-by: Mike Winter <mike.winter@digital.cabinet-office.gov.uk>